Kill mutation survivors by mirroring webhook setup tests
The mutation gate only runs a source's mirror-located direct tests plus
integration tests changed on the branch, so the endpoint-limit recovery
tests in test/integration/stripe/webhook-setup.test.ts never guarded
src/shared/stripe/endpoints.ts during mutation runs. They are fetch-mock
unit tests of that module, so they now live at its mirror path.

Also close the three remaining assertion gaps:
- cleanup asserts the Stripe client is created with zero network retries
- connection health asserts the listing error is reported word for word
- the webhook endpoint schema test accepts both endpoint statuses
